+++ crates/swc_webpack_ast/tests/fixture/array-literal/output.js
@@ -1,3 +1,0 @@
-import { a } from "x";
-import b from "y";
-const arr = (a, b());