+++ crates/swc_webpack_ast/tests/fixture/libs/1/output.js
@@ -1,1 +1,0 @@
-if (!require("./is-implemented")()) require("./shim");