+++ crates/swc_webpack_ast/tests/fixture/try-catch/1/output.js
@@ -1,5 +1,0 @@
-export function foo() {
-    const meta = null;
-    // const { event, team, user } = this.props;
-    const b = null;
-}